Query needs to select all notifications, not just unread ones
When we are querying the Github API to list notifications, we need to
query all notifications, not just unread ones. Otherwise, if a user
happens to, e.g., view the email for a notification (which
automatically causes it to be marked read in Github) in the interval
between two Gunsub scans, then we won't unsubscribe that notification
when we should.